hw-5: complete case 5

+++ src/main/kotlin/ru/quipy/common/utils/SlidingWindowRateLimiter.kt
@@ -9,17 +9,19 @@ import org.slf4j.LoggerFactory
 import java.time.Duration
 import java.util.concurrent.Executors
 import java.util.concurrent.PriorityBlockingQueue
+import java.util.concurrent.atomic.AtomicInteger
 import java.util.concurrent.atomic.AtomicLong
 import java.util.concurrent.locks.ReentrantLock
+import kotlin.concurrent.atomics.AtomicInt
 import kotlin.concurrent.withLock
 
 class SlidingWindowRateLimiter(
-    private val rate: Long,
+    private val rate: Int,
     private val window: Duration,
 ) : RateLimiter {
     private val rateLimiterScope = CoroutineScope(Executors.newSingleThreadExecutor().asCoroutineDispatcher())
 
-    private val sum = AtomicLong(0)
+    private val sum = AtomicInteger(0)
     private val queue = PriorityBlockingQueue<Measure>(10_000)
 
     override fun tick(): Boolean {
